Updated Item Bonuses Sytsem, Loader, Files, Integer input system added, String (long string) input system added, Safety Staff account login pin, Minor improvements

+++ 727 server/src/com/rs/game/dialogue/DialogueEventListener.java
@@ -23,7 +23,7 @@ public abstract class DialogueEventListener implements DialogueFaceExpression {
 	private Object[] args;
 	
 	public DialogueEventListener mes(String message, Object...args){
-		dialogueEvent.add(new DialogueEvent(0, String.format(message, args)));
+		dialogueEvent.add(new DialogueEvent((byte) 0, String.format(message, args)));
 		return this;
 	}
